Rename otp-digits OTPDigits, update buttercup
