+++ package.json
@@ -51,14 +51,14 @@
     "@babel/plugin-transform-spread": "^7.2.2",
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
-    "@buttercup/facades": "^0.7.0",
+    "@buttercup/facades": "^0.8.0",
     "@storybook/addon-actions": "^4.1.11",
     "@storybook/addons": "^4.1.11",
     "@storybook/react": "^4.1.11",
     "babel-loader": "^8.0.5",
     "babel-plugin-jsx-control-statements": "^4.0.0",
     "babel-plugin-ramda": "^2.0.0",
-    "buttercup": "^2.10.0",
+    "buttercup": "^2.13.0",
     "css-loader": "^2.1.0",
     "enzyme": "^3.8.0",
     "enzyme-adapter-react-16": "^1.9.1",
